requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
browsermob-proxy==0.8.0
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cowsay==6.1
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
demjson3==3.0.6
dnspython==2.6.1
email_validator==2.2.0
fake-useragent==1.4.0
fastapi==0.114.1
fastapi-cli==0.0.5
fonttools==4.49.0
gensim==4.3.2
geographiclib==1.52
geopy==2.2.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
paramiko==3.4.1
parse==1.20.1
pillow==10.2.0
pluggy==1.5.0
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.21
pydantic==2.9.1
pydantic_core==2.23.3
pyee==8.2.2
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.2
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
retry==0.9.2
rfc3986==1.5.0
rich==13.8.1
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.17.2
selenium-stealth==1.0.6
sentiment-analysis-spanish==0.0.25
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.38.5
threadpoolctl==3.4.0
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.30.6
uvloop==0.20.0
w3lib==2.1.2
watchfiles==0.24.0
webdriver-manager==4.0.2
websockets==10.4
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0
